requirements.txt
# ============================================================================
# AI Log Analyzer - Python Dependencies
# ============================================================================

# Database
psycopg2-binary>=2.9.0

# Environment & Config
python-dotenv>=1.0.0
PyYAML>=6.0

# HTTP requests (for ES)
requests>=2.28.0
urllib3>=1.26.0

# Data processing (optional)
# pandas>=1.5.0
# numpy>=1.23.0
